Gestión de pedidos y usuario administrador

+++ OfficePack/Pedido/views.py
@@ -3,29 +3,55 @@ from django.contrib.auth.decorators import login_required
 from django.core.mail import send_mail
 from django.conf import settings
 from .forms import PedidoForm
-from .models import Pedido, MetodoPago
+from .models import Pedido
 from Producto.models import Producto
 import stripe
 from django.http import JsonResponse
 from django.template.loader import render_to_string
+from django.http import HttpResponseForbidden
 
 stripe.api_key = settings.STRIPE_TEST_SECRET_KEY
 
 
+@login_required(login_url='/login/')
 def crear_pedido(request):
+    # comprobamos que el user sea el admin
+    if request.user.email != "officepack@gmail.com":
+        return HttpResponseForbidden("No tienes autoridad para realizar esta operación")
+    
     if request.method == 'POST':
         form = PedidoForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('listar_pedidos')
+            return redirect('/gestionar_pedidos/')
     else:
         form = PedidoForm()
     return render(request, 'crear_pedido.html', {'form': form})
 
+@login_required(login_url='/login/')
+def actualizar_pedido(request, pedido_id):
+    # comprobamos que el user sea el admin
+    if request.user.email != "officepack@gmail.com":
+        return HttpResponseForbidden("No tienes autoridad para realizar esta operación")
+    
+    if request.method == 'POST':
+        form = PedidoForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('/gestionar_pedidos/')
+    else:
+        form = PedidoForm(instance=get_object_or_404(Pedido, id=pedido_id))
+    return render(request, 'actualizar_pedido.html', {'form': form})
 
+
+@login_required(login_url='/login/')
 def listar_pedidos(request):
+    # comprobamos que el user sea el admin
+    if request.user.email != "officepack@gmail.com":
+        return HttpResponseForbidden("No tienes autoridad para realizar esta operación")
+    
     pedidos = Pedido.objects.all()
-    return render(request, 'listar_pedidos.html', {'pedido': pedidos})
+    return render(request, 'gestionar_pedidos.html', {'pedidos': pedidos})
 
 
 def mostrar_pedido(request, pedido_id):
@@ -39,12 +65,17 @@ def listar_mis_pedidos(request):
     return render(request, 'listar_pedidos.html', {'pedidos': pedidos})
 
 
-def eliminar_pedido(request, pk):
-    pedido = get_object_or_404(Pedido, pk=pk)
+@login_required(login_url='/login/')
+def eliminar_pedido(request, pedido_id):
+    # comprobamos que el user sea el admin
+    if request.user.email != "officepack@gmail.com":
+        return HttpResponseForbidden("No tienes autoridad para realizar esta operación")
+    
+    pedido = get_object_or_404(Pedido, id=pedido_id)
     if request.method == 'POST':
         pedido.delete()
-        return redirect('listar_pedidos')
-    return render(request, 'eliminar_pedido.html', {'pedido': pedido})
+        return redirect('/gestionar_pedidos')
+    return redirect(request.META.get('HTTP_REFERER', ''))
 
 
 # Operaciones con la cesta
@@ -124,6 +155,10 @@ def pagar(request):
     cesta = request.session.get('cesta', {})
     total = sum(item['precio'] * item['cantidad'] for item in cesta.values())
 
+    # Costes de envío
+    if total < 30:
+        total += 5
+
     # PaymentIntent con el total de la cesta
     intent = stripe.PaymentIntent.create(
          amount=int(total * 100),  # Cantidad en centavos
@@ -139,6 +174,10 @@ def pagar(request):
 def confirmar_pago(request):
     cesta = request.session.get('cesta', {})
     total = sum(item['precio'] * item['cantidad'] for item in cesta.values())
+
+    # Costes de envío
+    if total < 30:
+        total += 5
 
     if request.method == 'POST':
